+++ test/CMTATStandaloneSnapshot.test.js
@@ -9,7 +9,7 @@ const {
 } = require('../CMTAT/test/deploymentUtils')
 const { buildCmtatInitParams } = require('./helpers/cmtatInitParams')
 
-describe('CMTAT Snapshot Standalone', function () {
+describe('SnapshotEngine CMTAT Standalone', function () {
   beforeEach(async function () {
     this.snapshotAdminMode = 'access-control'
     const {
@@ -19,38 +19,16 @@ describe('CMTAT Snapshot Standalone', function () {
     } = buildCmtatInitParams()
 
     Object.assign(this, await loadFixture(fixture))
-    this.cmtat = await ethers.deployContract('CMTATStandaloneSnapshot', [
-      this.admin.address,
-      ERC20Attributes,
-      extraInformationAttributes,
-      engines
-    ]).catch(async (error) => {
-      if (!String(error).includes('code is too large')) {
-        throw error
-      }
-
-      this.cmtat = await deployCMTATStandaloneWithParameter(
-        this.deployerAddress.address,
-        this._.address,
+    this.cmtat = await ethers.deployContract(
+      'CMTATStandaloneInternalSnapshot',
+      [
         this.admin.address,
-        ERC20Attributes.name,
-        ERC20Attributes.symbol,
-        ERC20Attributes.decimalsIrrevocable,
-        extraInformationAttributes.tokenId,
-        extraInformationAttributes.terms,
-        extraInformationAttributes.information,
+        ERC20Attributes,
+        extraInformationAttributes,
         engines
-      )
-      this.transferEngineMock = await ethers.deployContract('SnapshotEngine', [
-        this.cmtat.target, this.admin
-      ])
-      await this.cmtat.connect(this.admin).setSnapshotEngine(this.transferEngineMock)
-      return this.cmtat
-    })
-
-    if (!this.transferEngineMock) {
-      this.transferEngineMock = this.cmtat
-    }
+      ]
+    )
+    this.transferEngineMock = this.cmtat
   })
 
   registerSnapshotCommonSuites()