+++ test/ERC20SnapshotModule.test.js
@@ -2,18 +2,17 @@ const {
   registerSnapshotCommonSuites
 } = require('./ERC20SnapshotModuleCommon/registerSnapshotCommonSuites')
 const {
-  deployCMTATStandalone,
   fixture,
   loadFixture
 } = require('../CMTAT/test/deploymentUtils')
+const { deployExternalSnapshotCmtat } = require('./helpers/deployExternalSnapshotCmtat')
 describe('Standard - ERC20SnapshotModule', function () {
   beforeEach(async function () {
     Object.assign(this, await loadFixture(fixture))
     this.snapshotAdminMode = 'access-control'
-    this.cmtat = await deployCMTATStandalone(
+    this.cmtat = await deployExternalSnapshotCmtat(
       this._.address,
-      this.admin.address,
-      this.deployerAddress.address
+      this.admin.address
     )
     this.transferEngineMock = await ethers.deployContract('SnapshotEngine', [
       this.cmtat.target, this.admin