+++ test/ERC20SnapshotModuleOwnable2Step.test.js
@@ -2,19 +2,18 @@ const {
   registerSnapshotCommonSuites
 } = require('./ERC20SnapshotModuleCommon/registerSnapshotCommonSuites')
 const {
-  deployCMTATStandalone,
   fixture,
   loadFixture
 } = require('../CMTAT/test/deploymentUtils')
+const { deployExternalSnapshotCmtat } = require('./helpers/deployExternalSnapshotCmtat')
 
 describe('Standard - ERC20SnapshotModule Ownable2Step', function () {
   beforeEach(async function () {
     Object.assign(this, await loadFixture(fixture))
     this.snapshotAdminMode = 'ownable-2step'
-    this.cmtat = await deployCMTATStandalone(
+    this.cmtat = await deployExternalSnapshotCmtat(
       this._.address,
-      this.admin.address,
-      this.deployerAddress.address
+      this.admin.address
     )
     this.transferEngineMock = await ethers.deployContract('SnapshotEngineOwnable2Step', [
       this.cmtat.target, this.admin